move compute filter to filter
